Add all instructions from specification
This ensures (hopefully) that all instructions and registers from the
latest version of the specification are present in syntax file,
including the privileged specification. This adds quite a few CSRs and
instructions for hypervisors and an extension or two for the standard
specification.

There are a few instructions and registers that were previously
supported by the plugin that are no longer in the latest version of the
specification. Those **should** still be supported unless I made a
mistake and accidentally deleted one.

Also, rather than trying to do a bunch of matching, this now uses the
`syntax keyword` for the vast majority of instructions and registers,
though for situations where there is a lot of variation, a `syntax
match` may still be used.

This shoudl ensure support for RV32I/RV64I/RV128I, the
M, A, F, D, Q, and C extensions, as well as Zicsr, counters, and
Zifencei.